WebDevelopmentLab03/requirements.txt
streamlit
google-generativeai
requests
pandas
altair